pyproject.toml
[project]
name = "customer-churn-prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow>=2.6.0",
    "apache-airflow-providers-standard>=1.0.0",
    "dagshub>=0.1.24",
    "dvc-s3>=2.0.0",
    "fastapi>=0.95.0",
    "joblib>=1.0.0",
    "matplotlib>=3.3.0",
    "mlflow>=1.20.0",
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=0.19.0",
    "scikit-learn>=1.0.0",
    "seaborn>=0.11.0",
    "skops>=0.11.0",
    "uvicorn>=0.21.0",
    "xgboost>=1.5.0",
    
]
